add delegate/undelegate actions to the transaction planner rpc

+++ crates/core/component/stake/src/action_handler/delegate.rs
@@ -75,9 +75,9 @@ impl ActionHandler for Delegate {
         // should give approximately the same results, they may not give
         // exactly the same results.
         let expected_delegation_amount =
-            next_rate_data.delegation_amount(d.unbonded_amount.value() as u64);
+            next_rate_data.delegation_amount(d.unbonded_amount.value());
 
-        if expected_delegation_amount != d.delegation_amount.value() as u64 {
+        if expected_delegation_amount != d.delegation_amount.value() {
             return Err(anyhow::anyhow!(
                     "given {} unbonded stake, expected {} delegation tokens but description produces {}",
                     d.unbonded_amount,
